LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/finished-promise-defers-cleanup-expected.txt


FAIL View transition cleanup is deferred until after the frame where `finished` promise resolves assert_equals: expected object "[object ViewTransition]" but got null
